refactor: 텍스트 하이라이트 구분하는 util 함수 분리
